평가메뉴(미평가 장소, 기존 평가 장소 조회)구현, 계정설정 구현
평가
- 미평가 장소 리스트 확인 및 미평가 장소 평가 기능 추가
- 기존 평가 목록 조회 기능 구현

설정
- 사용자가 가입시 입력한 계정 정보를 수정할 수 있도록 구현

+++ src/main/java/kr/ac/hansung/maldives/web/controller/LocationController.java
@@ -33,7 +33,7 @@ public class LocationController {
 	}
 
 	@RequestMapping("/detail")
-	public String locationDetail(@RequestParam("store_idx") int store_idx, Model model) {
+	public String locationDetail(@RequestParam("store_idx") Long store_idx, Model model) {
 		Store store = storeService.getStoreById(store_idx);
 		model.addAttribute("store", store);
 		return "location/detail";
